Stable codebase for Kotlin IoT android app
This is a Kotlin android app that I’m working on which will manage webcams over a network. The app will enable users to add and remove webcams as well as edit their settings. Additionally, with the aid of companion software written in python, the app will enable its users to view the streams of a remote webcam through the android app. The app utilizes google firebase for its login and database functionalities. As of 11/21/23, functionality-wise, this app is about 75% complete. Unfinished portions of the project include the following features: setting camera sensors and the server software for the webcam. The GUI, login, and database functionalities were coded using learning materials provided by Dr. Qu on Percipio.